refactor(hooks): extract navigation logic into useNavigation hook
